Added files from Thursday 9/8/22 class
The catTest.js program and the lizardJSON.json file are used for the .JSON format file assignment on Canvas. Students are to complete the catTest.js program so the lizardJSON.json file is read into using JSON.parse() an object variable "lizard" in the program, then displayed in the console to verify the parsing worked correctly.